refactor: validations utils로 이동
