+++ src/hooks/useDevice.js
